move join-pair into /domain-service
