+++ src/domain/pair/domain-service/join-pair.ts
@@ -1,12 +1,12 @@
-import { DomainService } from "../shared/domainService"
-import { Team } from "../team/team"
-import { Pair } from "./pair"
-import { Participant } from "../participant/participant"
-import { IPairRepository } from "./interface/pair-repository"
-import { PairFactory } from "./pair-factory"
-import { GetVacantPairList } from "./domain-service/get-vacant-pair-list"
-import { GetParentTeam } from "./domain-service/get-parent-team"
-import { GeneratePairName } from "./domain-service/generate-pair-name"
+import { DomainService } from "../../shared/domainService"
+import { Team } from "../../team/team"
+import { Pair } from "../pair"
+import { Participant } from "../../participant/participant"
+import { IPairRepository } from "../interface/pair-repository"
+import { PairFactory } from "../pair-factory"
+import { GetVacantPairList } from "./get-vacant-pair-list"
+import { GetParentTeam } from "./get-parent-team"
+import { GeneratePairName } from "./generate-pair-name"
 
 export class JoinPair extends DomainService<"join-pair"> {
   constructor(
