+++ src/domain/pair/domain-service/tests/join-pair.test.ts
@@ -1,23 +1,23 @@
 import { mock } from "jest-mock-extended"
 import { JoinPair } from "../join-pair"
 // entity
-import { TeamId } from "../../team/team-id"
-import { Team } from "../../team/team"
-import { TeamName } from "../../team/team-name"
-import { Pair } from "../pair"
-import { PairId } from "../pair-id"
-import { PairName } from "../pair-name"
-import { Participant } from "../../participant/participant"
-import { ParticipantId } from "../../participant/participant-id"
-import { ParticipantName } from "../../participant/participant-name"
-import { Email } from "../../participant/email"
+import { TeamId } from "../../../team/team-id"
+import { Team } from "../../../team/team"
+import { TeamName } from "../../../team/team-name"
+import { Pair } from "../../pair"
+import { PairId } from "../../pair-id"
+import { PairName } from "../../pair-name"
+import { Participant } from "../../../participant/participant"
+import { ParticipantId } from "../../../participant/participant-id"
+import { ParticipantName } from "../../../participant/participant-name"
+import { Email } from "../../../participant/email"
 // repository
-import { IPairRepository } from "../interface/pair-repository"
+import { IPairRepository } from "../../interface/pair-repository"
 // domain service
-import { PairFactory } from "../pair-factory"
-import { GetVacantPairList } from "../domain-service/get-vacant-pair-list"
-import { GetParentTeam } from "../domain-service/get-parent-team"
-import { GeneratePairName } from "../domain-service/generate-pair-name"
+import { PairFactory } from "../../pair-factory"
+import { GetVacantPairList } from "../get-vacant-pair-list"
+import { GetParentTeam } from "../get-parent-team"
+import { GeneratePairName } from "../generate-pair-name"
 
 describe("JoinPair", () => {
   const pairRepositoryMock = mock<IPairRepository>()
